The Great QEMU Rewrite of 2018
Create a hierarchical object model of QEMU's state which can be queried and
updated. Amongst other things, this allows us to use external snapshots, to
restart QEMU during the test and serialize our model of QEMU's
state (qemu_state.json).

Previously we used internal snapshots (i.e. snapshots which are saved inside an
existing qcow2 image), which are easy to use in simple cases. However very
little effort is put into maintaining these as the cloud providers use
external snapshots and migrations to file. Internal snapshots are known to
have bugs and performance issues which are not present when using external
snapshots.

Also when reverting to a snapshot, we previously did not restart QEMU. In
theory this should not be problem if every device in QEMU follows QEMU's
device model, however it is suspected that they do not. This means devices
could maintain state from before a snapshot is reverted creating an
inconsistent system state. Restarting QEMU ensures the devices are at least in
their default state before the snapshot is loaded.

In order to load and save the VM RAM and CPU state, we now use the migration
interface which is much better tested and maintained than the 'savevm' and
'loadvm' commands. However this makes loading a snapshot much more complex as
the user now needs to load the block device snapshots and RAM image
separately.

This also changes how UEFI pflash drives are used. It is now possible to
specify two pflash drives, one for the UEFI firmware ROM image and one for the
UEFI vars. This allows ARM VMs to be booted with no workarounds. This commit
also enables snapshots on ARM machines where they were previously disabled due
to being too unreliable.

Some features have been removed because they did not appear to have any users
nor did it seem possible to use them.

This is submitted as one very large commit because the code is highly
interdependent.

Signed-off-by: Ettore Di Giacinto <edigiacinto@suse.com>